Site/endpoint/service property delete checkboxes

+++ lib/Gocdb_Services/ServiceService.php
@@ -897,32 +897,69 @@ class ServiceService extends AbstractEntityService {
     }
 
     /**
+	 * DEPRECATED
      * Deletes a service property
      * @param \Service $service
      * @param \User $user
      * @param \SiteProperty $prop
      */
-    public function deleteServiceProperty(\Service $service, \User $user, \ServiceProperty $prop) {
-	//Check the portal is not in read only mode, throws exception if it is
-	$this->checkPortalIsNotReadOnlyOrUserIsAdmin($user);
-	$this->validateAddEditDeleteActions($user, $service);
+//    public function deleteServiceProperty(\Service $service, \User $user, \ServiceProperty $prop) {
+//	//Check the portal is not in read only mode, throws exception if it is
+//	$this->checkPortalIsNotReadOnlyOrUserIsAdmin($user);
+//	$this->validateAddEditDeleteActions($user, $service);
+//
+//	$this->em->getConnection()->beginTransaction();
+//	try {
+//	    // Service is the owning side so remove elements from service.
+//	    $service->getServiceProperties()->removeElement($prop);
+//	    // Once relationship is removed delete the actual element
+//	    $this->em->remove($prop);
+//	    $this->em->flush();
+//	    $this->em->getConnection()->commit();
+//	} catch (\Exception $e) {
+//	    $this->em->getConnection()->rollback();
+//	    $this->em->close();
+//	    throw $e;
+//	}
+//    }
 
-	$this->em->getConnection()->beginTransaction();
-	try {
-	    // Service is the owning side so remove elements from service.
-	    $service->getServiceProperties()->removeElement($prop);
-	    // Once relationship is removed delete the actual element
-	    $this->em->remove($prop);
-	    $this->em->flush();
-	    $this->em->getConnection()->commit();
-	} catch (\Exception $e) {
-	    $this->em->getConnection()->rollback();
-	    $this->em->close();
-	    throw $e;
+	/**
+	 * Deletes service properties
+	 * @param \Service $service
+	 * @param \User $user
+	 * @param array $propArr
+	 */
+	public function deleteServiceProperties(\Service $service, \User $user, array $propArr) {
+		//Check the portal is not in read only mode, throws exception if it is
+		$this->checkPortalIsNotReadOnlyOrUserIsAdmin($user);
+		$this->validateAddEditDeleteActions($user, $service);
+
+		$this->em->getConnection()->beginTransaction();
+		try {
+			foreach ($propArr as $prop) {
+				//throw new \Exception(var_dump($prop));
+				//check property is in service
+				if ($prop->getParentService() != $service){
+					$id = $prop->getId();
+					throw new \Exception("Property {$id} does not belong to the specified service");
+				}
+
+				// Service is the owning side so remove elements from service.
+				$service->getServiceProperties()->removeElement($prop);
+				// Once relationship is removed delete the actual element
+				$this->em->remove($prop);
+			}
+			$this->em->flush();
+			$this->em->getConnection()->commit();
+		} catch (\Exception $e) {
+			$this->em->getConnection()->rollback();
+			$this->em->close();
+			throw $e;
+		}
 	}
-    }
 
     /**
+	 * DEPRECATED
      * Deletes the given EndpointProperty from its parent Endpoint (if set).
      * If the parent Endpoint has not been set (<code>$prop->getParentEndpoint()</code> returns null) 
      * then the function simply returns because the user permissions to delete 
@@ -930,31 +967,73 @@ class ServiceService extends AbstractEntityService {
      * @param \User $user
      * @param \EndpointProperty $prop
      */
-    public function deleteEndpointProperty(\User $user, \EndpointProperty $prop) {
-	//Check the portal is not in read only mode, throws exception if it is
-	$this->checkPortalIsNotReadOnlyOrUserIsAdmin($user);
-	$endpoint = $prop->getParentEndpoint();
-	if ($endpoint == null) {
-	    // property endpoint hasn't been set, so just return. 
-	    return;
-	}
-	$service = $endpoint->getService();
-	$this->validateAddEditDeleteActions($user, $service);
+//    public function deleteEndpointProperty(\User $user, \EndpointProperty $prop) {
+//	//Check the portal is not in read only mode, throws exception if it is
+//	$this->checkPortalIsNotReadOnlyOrUserIsAdmin($user);
+//	$endpoint = $prop->getParentEndpoint();
+//	if ($endpoint == null) {
+//	    // property endpoint hasn't been set, so just return.
+//	    return;
+//	}
+//	$service = $endpoint->getService();
+//	$this->validateAddEditDeleteActions($user, $service);
+//
+//	$this->em->getConnection()->beginTransaction();
+//	try {
+//	    // EndointLocation is the owning side so remove elements from endpoint.
+//	    $endpoint->getEndpointProperties()->removeElement($prop);
+//	    // Once relationship is removed delete the actual element
+//	    $this->em->remove($prop);
+//	    $this->em->flush();
+//	    $this->em->getConnection()->commit();
+//	} catch (\Exception $e) {
+//	    $this->em->getConnection()->rollback();
+//	    $this->em->close();
+//	    throw $e;
+//	}
+//    }
 
-	$this->em->getConnection()->beginTransaction();
-	try {
-	    // EndointLocation is the owning side so remove elements from endpoint.
-	    $endpoint->getEndpointProperties()->removeElement($prop);
-	    // Once relationship is removed delete the actual element
-	    $this->em->remove($prop);
-	    $this->em->flush();
-	    $this->em->getConnection()->commit();
-	} catch (\Exception $e) {
-	    $this->em->getConnection()->rollback();
-	    $this->em->close();
-	    throw $e;
+	/**
+	 * Deletes the given EndpointProperties in the array from their parent Endpoints (if set).
+	 * If the parent Endpoint has not been set (<code>$prop->getParentEndpoint()</code> returns null
+	 * then the function throws an exception because the user permissions to delete
+	 * the EP can't be determined on a null Endpoint.
+	 * @param \User $user
+	 * @param array $propArr
+	 */
+	public function deleteEndpointProperties(\User $user, array $propArr) {
+		//Check the portal is not in read only mode, throws exception if it is
+		$this->checkPortalIsNotReadOnlyOrUserIsAdmin($user);
+
+		$this->em->getConnection()->beginTransaction();
+
+		try {
+			foreach ($propArr as $prop) {
+
+				//check endpoint property has an parent endpoint
+				$endpoint = $prop->getParentEndpoint();
+				if ($endpoint == null){
+					$id = $prop->getId();
+					throw new \Exception("Property {$id} does not have a parent endpoint");
+				}
+
+				//check user has permissions over the service associated with the endpoint
+				$service = $endpoint->getService();
+				$this->validateAddEditDeleteActions($user, $service);
+
+				// EndointLocation is the owning side so remove elements from endpoint.
+				$endpoint->getEndpointProperties()->removeElement($prop);
+				// Once relationship is removed delete the actual element
+				$this->em->remove($prop);
+			}
+			$this->em->flush();
+			$this->em->getConnection()->commit();
+		} catch (\Exception $e) {
+			$this->em->getConnection()->rollback();
+			$this->em->close();
+			throw $e;
+		}
 	}
-    }
 
     /**
      * Edits a service property
